LayoutTests/platform/wk2/storage/indexeddb/create-and-remove-object-store-expected.txt

Test IndexedDB's create and removeObjectStore

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
db.createObjectStore('tmp')
Expecting exception from db.createObjectStore('tmp')
PASS Exception was thrown.
PASS code is 0
PASS ename is 'ConstraintError'
Exception message: A mutation operation in the transaction failed because a constraint was not satisfied.
trans = db.transaction(['tmp'])
trans.objectStore('tmp').get(0)
PASS event.target.result is undefined.
Trying create
Expecting exception from db.createObjectStore("some os")
PASS Exception was thrown.
PASS code is DOMException.INVALID_STATE_ERR
PASS ename is 'InvalidStateError'
Exception message: An operation was called on an object on which it is not allowed or at a time when it is not allowed.
Trying remove
Expecting exception from db.deleteObjectStore("some os")
PASS Exception was thrown.
PASS code is DOMException.INVALID_STATE_ERR
PASS ename is 'InvalidStateError'
Exception message: An operation was called on an object on which it is not allowed or at a time when it is not allowed.
Trying create with store that already exists
Expecting exception from db.createObjectStore('tmp')
PASS Exception was thrown.
PASS code is DOMException.INVALID_STATE_ERR
PASS ename is 'InvalidStateError'
Exception message: An operation was called on an object on which it is not allowed or at a time when it is not allowed.
Trying remove with store that already exists
Expecting exception from db.deleteObjectStore('tmp')
PASS Exception was thrown.
PASS code is DOMException.INVALID_STATE_ERR
PASS ename is 'InvalidStateError'
Exception message: An operation was called on an object on which it is not allowed or at a time when it is not allowed.
PASS successfullyParsed is true

TEST COMPLETE
